+++ odoo/local-src/qoqa_offer_update/wizard/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+
+from . import qoqa_offer_update_wizard